+++ game_patch/os/win32_console.cpp
@@ -14,6 +14,10 @@
 
 static bool win32_console_enabled = false;
 static bool win32_console_input_line_printed = false;
+static HANDLE win32_console_input_handle;
+static HANDLE win32_console_output_handle;
+static bool win32_console_is_output_redirected = false;
+static bool win32_console_is_input_redirected = false;
 
 bool win32_console_is_enabled()
 {
@@ -22,31 +26,35 @@ bool win32_console_is_enabled()
 
 static void reset_console_cursor_column(bool clear)
 {
-    HANDLE output_handle = GetStdHandle(STD_OUTPUT_HANDLE);
+    if (win32_console_is_output_redirected) {
+        return;
+    }
     CONSOLE_SCREEN_BUFFER_INFO scr_buf_info;
-    GetConsoleScreenBufferInfo(output_handle, &scr_buf_info);
+    GetConsoleScreenBufferInfo(win32_console_output_handle, &scr_buf_info);
     if (scr_buf_info.dwCursorPosition.X == 0)
         return;
     COORD new_pos = scr_buf_info.dwCursorPosition;
     new_pos.X = 0;
-    SetConsoleCursorPosition(output_handle, new_pos);
+    SetConsoleCursorPosition(win32_console_output_handle, new_pos);
     if (clear) {
         for (int i = 0; i < scr_buf_info.dwCursorPosition.X; ++i) {
-            WriteConsoleA(output_handle, " ", 1, nullptr, nullptr);
+            WriteConsoleA(win32_console_output_handle, " ", 1, nullptr, nullptr);
         }
-        SetConsoleCursorPosition(output_handle, new_pos);
+        SetConsoleCursorPosition(win32_console_output_handle, new_pos);
         win32_console_input_line_printed = false;
     }
 }
 
 static void print_cmd_input_line()
 {
-    HANDLE output_handle = GetStdHandle(STD_OUTPUT_HANDLE);
+    if (win32_console_is_output_redirected) {
+        return;
+    }
     CONSOLE_SCREEN_BUFFER_INFO scr_buf_info;
-    GetConsoleScreenBufferInfo(output_handle, &scr_buf_info);
-    WriteConsoleA(output_handle, "] ", 2, nullptr, nullptr);
+    GetConsoleScreenBufferInfo(win32_console_output_handle, &scr_buf_info);
+    WriteConsoleA(win32_console_output_handle, "] ", 2, nullptr, nullptr);
     unsigned offset = std::max(0, rf::console::cmd_line_len - scr_buf_info.dwSize.X + 3);
-    WriteConsoleA(output_handle, rf::console::cmd_line + offset, rf::console::cmd_line_len - offset, nullptr, nullptr);
+    WriteConsoleA(win32_console_output_handle, rf::console::cmd_line + offset, rf::console::cmd_line_len - offset, nullptr, nullptr);
     win32_console_input_line_printed = true;
 }
 
@@ -86,23 +94,80 @@ void win32_console_init()
         return;
     }
 
-    AllocConsole();
+    win32_console_input_handle = GetStdHandle(STD_INPUT_HANDLE);
+    win32_console_output_handle = GetStdHandle(STD_OUTPUT_HANDLE);
+
+    char buf[256];
+    if (GetFinalPathNameByHandleA(win32_console_output_handle, buf, _countof(buf), 0) == 0) {
+        std::sprintf(buf, "(error %lu)", GetLastError());
+    }
+    xlog::info("Standard output info: path_name %s, file_type: %ld, handle %p",
+        buf, GetFileType(win32_console_output_handle), win32_console_output_handle);
+
+    if (!GetFileType(win32_console_output_handle)) {
+        if (!AllocConsole()) {
+            xlog::warn("AllocConsole failed, error %lu", GetLastError());
+        }
+        win32_console_input_handle = GetStdHandle(STD_INPUT_HANDLE);
+        win32_console_output_handle = GetStdHandle(STD_OUTPUT_HANDLE);
+        // We are using native console functions but in case some code tried stdio API reopen standard streams
+        std::freopen("CONOUT$", "w", stdout);
+        std::freopen("CONOUT$", "w", stderr);
+        std::freopen("CONIN$", "r", stdin);
+        xlog::info("Allocated new console, standard output: file_type %ld, handle %p",
+            GetFileType(win32_console_output_handle), win32_console_output_handle);
+    }
+
     SetConsoleCtrlHandler(console_ctrl_handler, TRUE);
+
     DWORD mode;
-    GetConsoleMode(GetStdHandle(STD_INPUT_HANDLE), &mode);
-    SetConsoleMode(GetStdHandle(STD_INPUT_HANDLE), mode & ~ ENABLE_ECHO_INPUT);
+    win32_console_is_output_redirected = !GetConsoleMode(win32_console_output_handle, &mode);
+    win32_console_is_input_redirected = !GetConsoleMode(win32_console_input_handle, &mode);
+    if (!win32_console_is_input_redirected) {
+        SetConsoleMode(win32_console_input_handle, mode & ~ ENABLE_ECHO_INPUT);
+    }
 
     // std::thread input_thread(input_thread_proc);
     // input_thread.detach();
 }
 
+static void write_console_output(std::string_view str)
+{
+    HANDLE output_handle = win32_console_output_handle;
+    if (!win32_console_is_output_redirected) {
+        WriteConsoleA(output_handle, str.data(), str.size(), NULL, NULL);
+    }
+    else {
+        // Convert LF to CRLF
+        DWORD bytes_written;
+        size_t start_pos = 0;
+        while (true) {
+            size_t end_pos = str.find('\n', start_pos);
+            if (end_pos == std::string_view::npos) {
+                WriteFile(output_handle, str.data() + start_pos, str.size() - start_pos, &bytes_written, NULL);
+                break;
+            }
+
+            size_t next_pos = end_pos + 1;
+            if (end_pos > 0 && str[end_pos - 1] == '\r') {
+                --end_pos;
+            }
+            if (end_pos - start_pos > 0) {
+                WriteFile(output_handle, str.data() + start_pos, end_pos - start_pos, &bytes_written, NULL);
+            }
+            WriteFile(output_handle, "\r\n", 2, &bytes_written, NULL);
+            start_pos = next_pos;
+        }
+    }
+}
+
 void win32_console_output(const char* text, [[maybe_unused]] const rf::Color* color)
 {
-    HANDLE output_handle = GetStdHandle(STD_OUTPUT_HANDLE);
     constexpr WORD red_attr = FOREGROUND_RED | FOREGROUND_INTENSITY;
     constexpr WORD blue_attr = FOREGROUND_BLUE | FOREGROUND_INTENSITY;
     constexpr WORD white_attr = FOREGROUND_RED | FOREGROUND_GREEN | FOREGROUND_BLUE | FOREGROUND_INTENSITY;
     constexpr WORD gray_attr = FOREGROUND_RED | FOREGROUND_GREEN | FOREGROUND_BLUE;
+    HANDLE output_handle = win32_console_output_handle;
     WORD current_attr = 0;
 
     if (win32_console_input_line_printed) {
@@ -142,26 +207,28 @@ void win32_console_output(const char* text, [[maybe_unused]] const rf::Color* co
             attr = gray_attr;
         }
 
-        if (current_attr != attr) {
+        if (current_attr != attr && !win32_console_is_output_redirected) {
             current_attr = attr;
             SetConsoleTextAttribute(output_handle, attr);
         }
 
-        WriteFile(output_handle, text_part.data(), text_part.size(), nullptr, nullptr);
+        write_console_output(text_part);
         pos = end_pos;
     }
 
-    if (text_sv.size() > 0 && text_sv[text_sv.size() - 1] != '\n')
-        WriteFile(output_handle, "\n", 1, nullptr, nullptr);
+    if (text_sv.size() > 0 && text_sv[text_sv.size() - 1] != '\n') {
+        write_console_output("\n");
+    }
 
-    if (current_attr != gray_attr)
+    if (current_attr != gray_attr && !win32_console_is_output_redirected) {
         SetConsoleTextAttribute(output_handle, gray_attr);
+    }
 }
 
 void win32_console_new_line()
 {
-    HANDLE output_handle = GetStdHandle(STD_OUTPUT_HANDLE);
-    WriteConsoleA(output_handle, "\r\n", 2, nullptr, nullptr);
+    DWORD bytes_written;
+    WriteFile(win32_console_output_handle, "\r\n", 2, &bytes_written, nullptr);
     print_cmd_input_line();
 }
 
@@ -177,7 +244,10 @@ void win32_console_update()
 
 void win32_console_poll_input()
 {
-    HANDLE input_handle = GetStdHandle(STD_INPUT_HANDLE);
+    if (win32_console_is_input_redirected) {
+        return;
+    }
+    HANDLE input_handle = win32_console_input_handle;
     INPUT_RECORD input_record;
     DWORD num_read = 0;
     while (true) {