+++ launcher_common/InjectingProcessLauncher.cpp
@@ -67,7 +67,7 @@ InjectingProcessLauncher::InjectingProcessLauncher(
     PROCESS_INFORMATION process_info;
     ZeroMemory(&process_info, sizeof(process_info));
     xlog::info("Calling CreateProcessA app_name %s, command_line %s, work_dir %s", app_name, command_line, work_dir);
-    BOOL result = CreateProcessA(app_name, const_cast<char*>(command_line), nullptr, nullptr, FALSE, CREATE_SUSPENDED,
+    BOOL result = CreateProcessA(app_name, const_cast<char*>(command_line), nullptr, nullptr, TRUE, CREATE_SUSPENDED,
                                  nullptr, work_dir, &startup_info, &process_info);
     xlog::info("CreateProcessA returned %d", result);
     if (!result) {